Rename project to AzurApi-Kotlin
