+++ README.md
@@ -1,2 +1,2 @@
-# Kongou
+# AzurApi-Kotlin
 A Kotlin wrapper for Azur Lane JSON
