+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'Kongou'
+rootProject.name = 'AzurApi-Kotlin'
