+++ src/main/kotlin/com/github/sylux6/azurapikotlin/api/AtagoInfo.kt
@@ -1,0 +1,8 @@
+package com.github.sylux6.azurapikotlin.api
+
+/**
+ * API info
+ */
+object AtagoInfo {
+    const val VERSION = "1.0.0"
+}
