+++ src/main/kotlin/com/github/sylux6/azurapikotlin/database/TakaoInfo.kt
@@ -1,9 +1,9 @@
-package com.github.sylux6.kongou.database
+package com.github.sylux6.azurapikotlin.database
 
 /**
  * JSON deserializer info
  */
-object HarunaInfo {
+object TakaoInfo {
     const val VERSION = "1.0.0"
     const val JSON_SOURCE = "https://raw.githubusercontent.com/jasperchua99/azur-json/master/ships/ships.json"
 }
