+++ src/main/kotlin/com/github/sylux6/azurapikotlin/internal/exceptions/ApiException.kt
@@ -1,0 +1,4 @@
+package com.github.sylux6.azurapikotlin.internal.exceptions
+
+class ApiException(s: String = "") : Exception(s) {
+}
