+++ src/main/kotlin/com/github/sylux6/azurapikotlin/internal/responses/ErrorResponse.kt
@@ -1,0 +1,4 @@
+package com.github.sylux6.azurapikotlin.internal.responses
+
+class ErrorResponse {
+}
