--- src/main/kotlin/com/github/sylux6/kongou/api/KongouInfo.kt
@@ -1,8 +1,0 @@
-package com.github.sylux6.kongou.api
-
-/**
- * API info
- */
-object KongouInfo {
-    const val VERSION = "1.0.0"
-}
