--- src/main/kotlin/com/github/sylux6/kongou/internal/exceptions/ApiException.kt
@@ -1,4 +1,0 @@
-package com.github.sylux6.kongou.internal.exceptions
-
-class ApiException(s: String = "") : Exception(s) {
-}
