--- src/main/kotlin/com/github/sylux6/kongou/internal/responses/ErrorResponse.kt
@@ -1,4 +1,0 @@
-package com.github.sylux6.kongou.internal.responses
-
-class ErrorResponse {
-}
