+++ src/test/kotlin/com/github/sylux6/azurapikotlin/api/AtagoTests.kt
@@ -1,0 +1,5 @@
+package com.github.sylux6.azurapikotlin.api
+
+class AtagoTests {
+
+}
