--- src/test/kotlin/com/github/sylux6/kongou/api/KongouTests.kt
@@ -1,5 +1,0 @@
-package com.github.sylux6.kongou.api
-
-class KongouTests {
-
-}
